Making Sparkle 64-bit compatible, thanks to Nick Zitzmann

+++ NSFileManager+Authentication.m
@@ -90,7 +90,7 @@
 {
 	if ([[NSFileManager defaultManager] isWritableFileAtPath:dst] && [[NSFileManager defaultManager] isWritableFileAtPath:[dst stringByDeletingLastPathComponent]])
 	{
-		int tag = 0;
+		NSInteger tag = 0;
 		BOOL result = [[NSWorkspace sharedWorkspace] performFileOperation:NSWorkspaceRecycleOperation source:[dst stringByDeletingLastPathComponent] destination:@"" files:[NSArray arrayWithObject:[dst lastPathComponent]] tag:&tag];
 		result &= [[NSFileManager defaultManager] copyPath:src toPath:dst handler:nil];
 		return result;
